ref : split PantryBox code
